\name{svDoc-package}
\alias{svDoc-package}
\alias{svDoc}
\docType{package}

\title{ SciViews GUI API - SciVews document management }
\description{
  The SciViews svDoc package provides additional function to combine Sweave,
  Asciidoc and Shiny to make flexible and dynamic documents.
}
\details{
  \tabular{ll}{
    Package: \tab svDoc\cr
    Type: \tab Package\cr
    Version: \tab 0.9-10\cr
    Date: \tab 2014-03-03\cr
    License: \tab GPL 2 or above, at your convenience\cr
  }
  % TODO: add description of main functions here. Also add examples
}

\author{
  Philippe Grosjean

  Maintainer: Ph. Grosjean <phgrosjean@sciviews.org>
}

\keyword{ package }
\keyword{ utilities }

\concept{ GUI API sweave Asciidoc litterate programming with Shiny widgets }
